refactor xml parser name helpers
